# frontend/kucukdevapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "kucukdevapi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/universities/{unid}/departments/{depid}/curriculums": {
      "get": {
        "tags": [
          "curriculums"
        ],
        "summary": "List Department Curriculums",
        "description": "list all curriculums of a department of a university with given universityID and universityDepartmentID",
        "operationId": "listUniversityDepartmentCurriculums",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Depid",
              "type": "string"
            },
            "name": "depid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "List all university curriculums",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Department Curriculums Universities  Unid  Departments  Depid  Curriculums Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UniversityCurriculumModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "curriculums"
        ],
        "summary": "Create Department Curriculum",
        "description": "Create department for a university with given universityID and universityDepartmentID",
        "operationId": "createUniversityDepartmentCurriculum",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Depid",
              "type": "string"
            },
            "name": "depid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityCurriculumModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Add new university department curriculum",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/universities/{unid}/departments/{depid}/curriculums/{curid}": {
      "get": {
        "tags": [
          "curriculums"
        ],
        "summary": "Show Department Curriculum",
        "description": "Get a single semester of a university with given universityID, universityDepartmentID and departmentCurriculumID",
        "operationId": "getSingleUniversityDepartmentCurriculum",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Depid",
              "type": "string"
            },
            "name": "depid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Curid",
              "type": "string"
            },
            "name": "curid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Show a university department curriculum",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityCurriculumModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "curriculums"
        ],
        "summary": "Update Department Curriculum",
        "description": "Update department of a university with given universityID, universityDepartmentID and departmentCurriculumID",
        "operationId": "updateUniversityDepartmentCurriculum",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Depid",
              "type": "string"
            },
            "name": "depid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Curid",
              "type": "string"
            },
            "name": "curid",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UniversityCurriculumModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update a university department curriculum",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "curriculums"
        ],
        "summary": "Delete Department Curriculum",
        "description": "Delete a university department with given universityID, universityDepartmentID and departmentCurriculumID",
        "operationId": "deleteUniversityDepartmentCurriculum",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Unid",
              "type": "string"
            },
            "name": "unid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Depid",
              "type": "string"
            },
            "name": "depid",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Curid",
              "type": "string"
            },
            "name": "curid",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete university department curriculum",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurriculumLessonModel": {
        "title": "CurriculumLessonModel",
        "required": [
          "name",
          "code",
          "lessonType"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "lessonType": {
            "title": "Lessontype",
            "type": "string"
          }
        },
        "example": {
          "name": "PHYSICS I",
          "code": "PHYS101",
          "lessonType": "science"
        }
      },
      "CurriculumSemesterModel": {
        "title": "CurriculumSemesterModel",
        "required": [
          "semester"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "semester": {
            "title": "Semester",
            "exclusiveMaximum": 9,
            "exclusiveMinimum": 0,
            "type": "integer"
          },
          "lessons": {
            "title": "Lessons",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurriculumLessonModel"
            },
            "default": []
          }
        },
        "example": {
          "semester": 1
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "Message": {
        "title": "Message",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "MessageCreate": {
        "title": "MessageCreate",
        "required": [
          "_id",
          "message"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "example": {
          "_id": "61ddea901311ecaed99afb7c",
          "message": "string"
        }
      },
...
      "UniversityCurriculumModel": {
        "title": "UniversityCurriculumModel",
        "required": [
          "name",
          "startYear",
          "endYear"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "startYear": {
            "title": "Startyear",
            "type": "integer"
          },
          "endYear": {
            "title": "Endyear",
            "type": "integer"
          },
          "semesters": {
            "title": "Semesters",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurriculumSemesterModel"
            },
            "default": []
          }
        },
        "example": {
          "name": "2016 Later",
          "startYear": 2016,
          "endYear": 2100
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}